Draft.

+++ _lib/kit/install/cloud/cloud-init/zRotten/zMinimal/rotten_install.sh
@@ -888,38 +888,44 @@ _install() {
 		#rm -f ./ubiquitous_bash.sh > /dev/null 2>&1
 		cp -f /ubiquitous_bash.sh ./ubiquitous_bash.sh
 	fi
-	chmod 755 ubiquitous_bash.sh
-	! ./ubiquitous_bash.sh _true && echo 'missing: ubiquitous_bash.sh' && exit 1
-	./ubiquitous_bash.sh _false && echo 'missing: ubiquitous_bash.sh' && exit 1
-	#"$scriptAbsoluteLocation" _setupUbiquitous
-	#"$scriptAbsoluteLocation" _getMost_debian11
-	#"$scriptAbsoluteLocation" _get_veracrypt
-	#"$scriptAbsoluteLocation" _test
+	if [[ -e ./ubiquitous_bash.sh ]]
+	then
+		chmod 755 ubiquitous_bash.sh
+		! ./ubiquitous_bash.sh _true && echo 'missing: ubiquitous_bash.sh' && exit 1
+		./ubiquitous_bash.sh _false && echo 'missing: ubiquitous_bash.sh' && exit 1
+		#"$scriptAbsoluteLocation" _setupUbiquitous
+		#"$scriptAbsoluteLocation" _getMost_debian11
+		#"$scriptAbsoluteLocation" _get_veracrypt
+		#"$scriptAbsoluteLocation" _test
+		
+		_install_and_run_package git
+		
+		sudo -n cp ./ubiquitous_bash.sh /home/"$custom_user"
+		sudo -n chown "user:user" "/home/""$custom_user""/ubiquitous_bash.sh"
+		sudo -n chmod "755" "/home/""$custom_user""/ubiquitous_bash.sh"
+		sudo -n chown "$custom_user":"$custom_user" "/home/""$custom_user""/ubiquitous_bash.sh"
+		#sudo -n -u user "/home/""$custom_user""/ubiquitous_bash.sh" _setupUbiquitous
+		sudo -n -u user sh -c "cd ; ""/home/""$custom_user""/ubiquitous_bash.sh"" _setupUbiquitous 2>&1"
+		
+		sudo -n cp ./ubiquitous_bash.sh /root
+		sudo -n chown "user:user" /root"/ubiquitous_bash.sh"
+		sudo -n chmod "755" /root"/ubiquitous_bash.sh"
+		sudo -n chown root:root /root"/ubiquitous_bash.sh"
+		#sudo -u root /root"/ubiquitous_bash.sh" _setupUbiquitous
+		sudo -u root sh -c "cd ; /root/ubiquitous_bash.sh _setupUbiquitous 2>&1"
+		
+		
+		
+		# ATTENTION: Optional. Attempts '_getMost' and '_test' from full 'ubiquitous_bash.sh' from upstream.
+		echo '________________________________________'
+		sudo -u root INSTANCE_ID="$INSTANCE_ID" sh -c "cd ; /root/bin/ubiquitous_bash.sh _getMost_debian11 2>&1"
+		sudo -u root INSTANCE_ID="$INSTANCE_ID" sh -c "cd ; /root/bin/ubiquitous_bash.sh _get_veracrypt 2>&1"
+		sudo -u root INSTANCE_ID="$INSTANCE_ID" sh -c "cd ; /root/bin/ubiquitous_bash.sh _test 2>&1" | tee /var/log/ubiquitous_bash-test
+		[[ ${PIPESTATUS[0]} != "0" ]] && _messageFAIL
+	fi
 	
-	_install_and_run_package git
-	
-	sudo -n cp ./ubiquitous_bash.sh /home/"$custom_user"
-	sudo -n chown "user:user" "/home/""$custom_user""/ubiquitous_bash.sh"
-	sudo -n chmod "755" "/home/""$custom_user""/ubiquitous_bash.sh"
-	sudo -n chown "$custom_user":"$custom_user" "/home/""$custom_user""/ubiquitous_bash.sh"
-	#sudo -n -u user "/home/""$custom_user""/ubiquitous_bash.sh" _setupUbiquitous
-	sudo -n -u user sh -c "cd ; ""/home/""$custom_user""/ubiquitous_bash.sh"" _setupUbiquitous 2>&1"
-	
-	sudo -n cp ./ubiquitous_bash.sh /root
-	sudo -n chown "user:user" /root"/ubiquitous_bash.sh"
-	sudo -n chmod "755" /root"/ubiquitous_bash.sh"
-	sudo -n chown root:root /root"/ubiquitous_bash.sh"
-	#sudo -u root /root"/ubiquitous_bash.sh" _setupUbiquitous
-	sudo -u root sh -c "cd ; /root/ubiquitous_bash.sh _setupUbiquitous 2>&1"
-	
-	
-	
-	# ATTENTION: Optional. Attempts '_getMost' and '_test' from full 'ubiquitous_bash.sh' from upstream.
 	echo '________________________________________'
-	sudo -u root INSTANCE_ID="$INSTANCE_ID" sh -c "cd ; /root/bin/ubiquitous_bash.sh _getMost_debian11 2>&1"
-	sudo -u root INSTANCE_ID="$INSTANCE_ID" sh -c "cd ; /root/bin/ubiquitous_bash.sh _get_veracrypt 2>&1"
-	sudo -u root INSTANCE_ID="$INSTANCE_ID" sh -c "cd ; /root/bin/ubiquitous_bash.sh _test 2>&1" | tee /var/log/ubiquitous_bash-test
-	[[ ${PIPESTATUS[0]} != "0" ]] && _messageFAIL
+	
 	sudo -n env DEBIAN_FRONTEND=noninteractive apt-get -o Dpkg::Options::="--force-confdef" -o Dpkg::Options::="--force-confold" -y upgrade
 	
 	sudo -n -u user bash -c "cd ; mkdir -p "/home/"$custom_user""/Downloads"
